Přejmenování rozhrani "Packetable"  podle konvencí javy na "IPacketable" Odstranění rozhraní "IPacketable" z třídy "Scheme" Vytvoření "SchemePacketHandleru", který je zodpovědný za převedení packetu na byty. Tím se zajistí jednodušší možnost modifikace. Implementace "SchemePacketHandleru" v ERPFragmentu
