+++ app/src/main/java/cz/zcu/fav/tymsnu/stimulatorremotecontrol/bytes/IPacketable.java
@@ -7,8 +7,8 @@ import java.util.ArrayList;
  * Interface, který dovoluje třídám jenž jej implementují umět sami sebe reprezentovat jako list packetů,
  * které můžeme následně poslat do stimulátoru.
  */
-public interface Packetable {
+public interface IPacketable {
 
-    public ArrayList<Packet> getPackets();
+    ArrayList<Packet> getPackets();
 
 }
